added a daily budget component and a component for my page buttons. also added in the styles for  income and expense for the add item container

+++ src/js/components/dumb/bottomSection/ListItem.js
@@ -7,7 +7,7 @@ const ListItem = props => {
             <div className="item clearfix" id={props.data.id}>
                 <div className="item__description">{props.data.description}</div>
                 <div className="right clearfix">
-                    <div className="item__value">{parseInt(props.data.value).toLocaleString("en-US", { style: "currency", currency: "USD" })}</div>
+                    <div className="item__value">{parseFloat(props.data.value).toLocaleString("en-US", { style: "currency", currency: "USD" })}</div>
                     {props.data.type === "exp" ? <div className="item__percentage">{props.totalIncome > 0 ? Math.round(parseInt(props.data.value) / props.totalIncome * 100) : "---"}</div>: ""}
                     <div className="item__delete">
                         <button className="item__delete--btn" onClick={() => props.handleDelete(props.data.id)}>x</button>
